# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rahul Raj — Portfolio & Systems Architecture API",
    "version": "1.0.0",
    "description": "Production-grade RESTful API documentation for Rahul Raj's modern portfolio platform, built with **Next.js 16**, **TypeScript**, **MongoDB Atlas**, **Cloudinary**, and **AI LLM Orchestration**.\n\n### Key Architectural Pillars\n- **Public Gateway**: Ultra-fast endpoints for portfolio contents, live ATS resume downloads, real-time presence heartbeats, and AI assistant interaction.\n- **Admin Control Studio & Hardware Biometrics**: Role-Based Access Control (RBAC) secured via WebAuthn/FIDO2 hardware biometric passkeys (Fingerprint, TouchID, Windows Hello, FaceID), 4-digit security PIN, master passcode, cryptographic HMAC tokens, or Email OTP resets.\n- **Resilience Engine**: Zero-crash persistence architecture backed by MongoDB Atlas with in-memory caching and real-time health telemetry probes.\n- **Security & Anti-Abuse**: IP-based sliding window rate limiters, anti-bot honeypot traps, and encrypted session cookies.",
    "contact": {
      "name": "Rahul Raj Modi",
      "email": "rahulraj21480@gmail.com",
      "url": "https://github.com/Rahul-2148"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment / Origin Server"
    },
    {
      "url": "http://localhost:3000",
...
    }
  ],
...
  "paths": {
...
    "/api/admin/skills": {
      "get": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "List technical and operational skills",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "skills": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Skill"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Create or update skill items",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Skill saved"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/resumes": {
      "get": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "List resumes with active switch status",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of resumes"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Add or set active primary resume",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resume updated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/portfolio": {
      "get": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Fetch live MongoDB Atlas portfolio dataset with connectivity health",
        "description": "Retrieves the master portfolio document from MongoDB Atlas or fallback static defaults if the cloud database is offline.",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio data and database sync status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isConnected": {
                      "type": "boolean",
                      "example": true
                    },
                    "isSynced": {
                      "type": "boolean",
                      "example": true
                    },
                    "source": {
                      "type": "string",
                      "example": "mongodb_atlas"
                    },
                    "message": {
                      "type": "string",
                      "example": "Connected to MongoDB Atlas. Live data loaded."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Admin CMS Content"
        ],
        "summary": "Sync or update full portfolio content",
        "description": "Performs full document saves, section-level updates, or seeds MongoDB Atlas with the codebase default data.",
        "security": [
          {
            "AdminCookieAuth": []
          },
          {
            "AdminPasscodeAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "sync_initial",
                      "update_section"
                    ],
                    "example": "sync_initial"
                  },
                  "section": {
                    "type": "string",
                    "example": "personalInfo"
                  },
                  "data": {
                    "type": "object",
                    "description": "Arbitrary section data or complete portfolio payload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changes successfully persisted to MongoDB Atlas"
          },
          "400": {
            "description": "Invalid payload or database disconnected"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Database write failure"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AdminCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "portfolio_admin_token",
        "description": "Encrypted HTTP-only admin session token generated during login"
      },
      "AdminPasscodeAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-passcode",
        "description": "Secret master passcode configured via ADMIN_PASSCODE environment variable"
      }
    },
    "schemas": {
...
      "Skill": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Next.js"
          },
          "category": {
            "type": "string",
            "example": "Frontend"
          },
          "level": {
            "type": "string",
            "enum": [
              "core",
              "proficient",
              "familiar"
            ],
            "example": "core"
          },
          "domain": {
            "type": "string",
            "enum": [
              "IT",
              "Non-IT"
            ],
            "example": "IT"
          }
        }
      },
...
    }
  }
}